% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCalibrationInfo.R
\name{addCalibrationInfo}
\alias{addCalibrationInfo}
\title{Add information about calibration.}
\usage{
addCalibrationInfo(MicroPEMObject, fileCalibration)
}
\arguments{
\item{MicroPEMObject}{the MicroPEM object}

\item{fileCalibration}{the path to the file where the information about calibration is stored}
}
\value{
A MicroPEMObject with calibration info.
}
\description{
Add information about calibration.
}

