% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarmCHAI.R
\name{alarmCHAI}
\alias{alarmCHAI}
\title{Generates alarm indicators for a MicroPEMObject.}
\usage{
alarmCHAI(MicroPEMObject)
}
\arguments{
\item{MicroPEMObject}{the MicroPEM object}
}
\value{
A \code{data table} with a column for the name of the indicator and a column with booleans. If no
alarm was flagged, the \code{data table} only has one line indicating that all is good.
}
\description{
Generates alarm indicators for a MicroPEMObject.
}
\examples{
data(dummyMicroPEMChai)
alarmCHAI(dummyMicroPEMChai)
}

