% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchConvert.R
\name{batchConvert}
\alias{batchConvert}
\title{Reading several MicroPEM files and converting the settings and measurement tables to csv.}
\usage{
batchConvert(path_input, path_output = path_input)
}
\arguments{
\item{path_input}{path to the directory with files}

\item{path_output}{path where the files should be created, by default equal to \code{path_input}.
See the data in inst/data to see
which one applies.

The function saves results in the input directory as csv files with a "," as separator.
One file is settings.csv with all settings, the other one is measures.csv
It saves them directly for not loading all of them at the same time in the session.}
}
\description{
Reading several MicroPEM files and converting the settings and measurement tables to csv.
}

