% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{CHAI}
\alias{CHAI}
\title{A  micropem output file.}
\format{A  csv micropem output file of version CHAI of the output file.}
\description{
A  csv micropem output file of version CHAI of the output file.
}
\examples{
\dontrun{
micropemExample <- convert_output(system.file('extdata', 'CHAI.csv', package = 'ammon'),
version='CHAI')
micropemExample$plot()
}
}
\keyword{data}

