% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chai_alarm.R
\name{chai_alarm}
\alias{chai_alarm}
\title{Generates alarm indicators for a micropem object.}
\usage{
chai_alarm(micropem_object)
}
\arguments{
\item{micropem_object}{the MicroPEM object}
}
\value{
A \code{data table} with a column for the name of the indicator and a column with booleans. If no
alarm was flagged, the \code{data table} only has one line indicating that all is good.
}
\description{
Generates alarm indicators for a micropem object.
}
\examples{
data(micropemChai)
chai_alarm(micropemChai)
}
