% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{Columbia2}
\alias{Columbia2}
\title{A  micropem output file.}
\format{A  csv micropem output file of version Columbia2 of the output file.}
\description{
A  csv micropem output file of version Columbia2 of the output file.
}
\examples{
\dontrun{
micropemExample <- convert_output(system.file('extdata', 'Columbia2.csv', package = 'ammon'),
version='Columbia2')
micropemExample$plot()
}
}
\keyword{data}

