% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSettings.R
\name{compareSettings}
\alias{compareSettings}
\title{Extracts parameters settings and some other control variables from the top of the output files of a given directory.}
\usage{
compareSettings(directory, version)
}
\arguments{
\item{directory}{the directory with all output files, of the same version.}

\item{version}{the version of the output files, either 'CHAI' or 'Columbia'}
}
\value{
A data frame tbl (class defined in the \code{dplyr} package).
}
\description{
Extracts parameters settings and some other control variables from the top of the output files of a given directory.
}

