% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_output.R
\name{convert_output}
\alias{convert_output}
\title{Uses output file from MicroPEM to create a MicroPEM object.}
\usage{
convert_output(path)
}
\arguments{
\item{path}{the path to the file}
}
\value{
A \code{MicroPEM} object.
}
\description{
Uses output file from MicroPEM to create a MicroPEM object.
}
\examples{
MicroPEMExample <- convert_output(system.file('extdata', 'dummyCHAI.csv', package = 'ammon'))
MicroPEMExample$plot()
}

