% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOutput.R
\name{convertOutput}
\alias{convertOutput}
\title{Uses output file from MicroPEM to create a MicroPEM object.}
\usage{
convertOutput(path)
}
\arguments{
\item{path}{the path to the file}
}
\value{
A \code{MicroPEM} object.
}
\description{
Uses output file from MicroPEM to create a MicroPEM object.
}
\examples{
MicroPEMExample <- convertOutput(system.file('extdata', 'dummyCHAI.csv', package = 'ammon'),
 version='CHAI')
MicroPEMExample$plot()
}

