% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dummyCHAI}
\alias{dummyCHAI}
\title{A dummy MicroPEM output file.}
\format{A dummy csv MicroPEM output file of version CHAI of the output file.}
\description{
A dummy csv MicroPEM output file of version CHAI of the output file.
}
\examples{
\dontrun{
MicroPEMExample <- convertOutput(system.file('extdata', 'dummyCHAI.csv', package = 'ammon'),
version='CHAI')
MicroPEMExample$plot()
}
}
\keyword{data}

