% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dummyColumbia}
\alias{dummyColumbia}
\title{A dummy MicroPEM output file.}
\format{A dummy csv MicroPEM output file of version Columbia of the output file.}
\description{
A dummy csv MicroPEM output file of version Columbia of the output file.
}
\examples{
\dontrun{
MicroPEMExample <- convert_output(system.file('extdata', 'dummyColumbia.csv', package = 'ammon'),
version='Columbia1')
MicroPEMExample$plot()
}
}
\keyword{data}

