% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dummyMicroPEMC1}
\alias{dummyMicroPEMC1}
\title{A dummy MicroPEM object.}
\format{A dummy MicroPEM object.}
\description{
A dummy MicroPEM object.
}
\examples{
data("dummyMicroPEMC1")
dummyMicroPEMC1$plot()
}
\keyword{data}

