% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dummyMicroPEMC2}
\alias{dummyMicroPEMC2}
\title{A dummy MicroPEM object.}
\format{A dummy MicroPEM object.}
\description{
A dummy MicroPEM object.
}
\examples{
data("dummyMicroPEMC2")
dummyMicroPEMC2$plot()
}
\keyword{data}

