% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dummyMicroPEMChai}
\alias{dummyMicroPEMChai}
\title{A dummy MicroPEM object.}
\format{A dummy MicroPEM object.}
\description{
A dummy MicroPEM object.
}
\examples{
data("dummyMicroPEMChai")
dummyMicroPEMChai$plot()
}
\keyword{data}

