% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTimeDate.R
\name{filterTimeDate}
\alias{filterTimeDate}
\title{Filter measures of the MicroPEM object in a given time interval.}
\usage{
filterTimeDate(MicroPEMObject = MicroPEMObject, fromTime = NULL,
  untilTime = NULL)
}
\arguments{
\item{MicroPEMObject}{A MicroPEMObject object}

\item{fromTime}{A POSIXct object defining the smallest observation timepoint}

\item{untilTime}{A POSIXct object defining the biggest observation timepoint}
}
\value{
A \code{MicroPEM} object.
}
\description{
Filter measures of the MicroPEM object in a given time interval.
}
\examples{
# load the lubridate package
library('lubridate')
# load the dummy MicroPEM object
data('dummyMicroPEMChai')
# look at the dimensions of the data.frame
print(dummyMicroPEMChai$measures)
# command for erasing measures from the first twelve hours
shorterMicroPEM <- filterTimeDate(MicroPEMObject=dummyMicroPEMChai,
untilTime=NULL,
fromTime=min(dummyMicroPEMChai$measures$timeDate, na.rm=TRUE) + hours(12))
# look at the dimensions of the data.frame
print(shorterMicroPEM$measures)
}

