% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammon.R
\docType{package}
\name{micropem-package}
\alias{micropem}
\alias{micropem-package}
\title{Supports the analyses of RTI MicroPEM output files.}
\description{
ammon allows you to read and analyse data from MicroPEM output files
in a reproducible way. MicroPEM are devices developped by RTI \url{http://www.rti.org/page.cfm/Aerosol_Sensors}.
It is called ammon like Zeus Ammon (\url{https://en.wikipedia.org/wiki/Amun#Greece})
because it helps us to Analyse Micropem MONitoring data
and it should be cool like a god. That being said, please don't worship it
and help us look for bugs in the code.
}

