% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microPEM-class.R
\docType{class}
\name{MicroPEM}
\alias{MicroPEM}
\title{An R6 class to represent MicroPEM output information.}
\format{\code{\link{R6Class}} object.}
\usage{
MicroPEM
}
\value{
Object of \code{\link{R6Class}}.
}
\description{
An R6 class to represent MicroPEM output information.
}
\section{Fields}{

\describe{
\item{\code{control}}{Data.frame (\code{dplyr "tbl_df"}) with settings of the MicroPEM device and other information such as download time.}

\item{\code{calibration}}{List of calibration information.}

\item{\code{measures}}{Data.frame (\code{dplyr "tbl_df"}) with all time-varying measures, possibly:
\describe{
  \item{datetime}{Time and date of each measurement, as a POSIXt object. Depending on the different logs of the time-varying variables there is not a measure for all variables associated to each timepoint.}
  \item{rh_corrected_nephelometer}{Measures of nephelometer in microgram/meter cube (numeric).}
  \item{temp}{Measures of temperature in centigrade (numeric).}
  \item{rh}{Measures of relative humidity that are a proportion and as such do not have an unit (numeric).}
  \item{battery}{Measures of battery  UNIT PLEASE (numeric).}
  \item{orifice_press}{Measures of orifice pressure in inches of water (numeric).}
  \item{inlet_press}{Measures of inlet pressure in inches of water (numeric).}
  \item{flow}{Measures of flow in liters per minute (numeric).}
  \item{x_axis}{x-axis accelerometer UNIT PLEASE (numeric).}
  \item{y_axis}{y-axis accelerometer UNIT PLEASE (numeric).}
  \item{z_axis}{z-axis accelerometer UNIT PLEASE (numeric).}
  \item{vector_sum_composite}{vector sum UNIT PLEASE (numeric).}
  \item{message}{Shutdown reason (factor).}
  }}

\item{\code{original}}{Boolean. Is this an original MicroPEM object (TRUE) or was it e.g. filtered or cleaned (FALSE).}
}}
\section{Methods}{

\describe{
  \item{plot}{Method for getting a quick plot of all time-varying measurements.
  Either \code{type ="plain"} or \code{type ="interactive"}, see examples.
  The method returns a plot of the \code{ggplot}-class.}
  \item{summary}{Method for getting a summary table (\code{dplyr "tbl_df"}) of all time-varying numeric measurements.}
  \item{print}{Method for printing both the summary table of all time-varying numeric measurements and all settings from the \code{control} field.}
  }
}
\examples{
data("dummyMicroPEMChai")
# Plot method, type = "plain" by default.
dummyMicroPEMChai$plot()
# Example with type = "interactive", for RStudio viewer,
# RMardown html documents and Shiny apps.
\dontrun{
library("ggiraph")
p <- dummyMicroPEMChai$plot(type = "interactive")
ggiraph(code = {print(p)}, width = 10, height = 10)}
# Summary method
dummyMicroPEMChai$summary()
# Print method
dummyMicroPEMChai$print()
}
\keyword{data}

