% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run a built-in Shiny App.}
\usage{
run_shiny_app()
}
\value{
Nothing, it does something.
}
\description{
Run a built-in Shiny App.
}
\examples{
\dontrun{
run_shiny_app()
}
}
