% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchConvert.R
\name{batchConvert}
\alias{batchConvert}
\title{Reading several MicroPEM files and converting the settings and measurement tables to csv.}
\usage{
batchConvert(pathDir, version = NULL)
}
\arguments{
\item{pathDir}{path to the directory with files}

\item{version}{the versions of the output file, either 'CHAI', 'Columbia1' or 'Columbia2'.
See the data in inst/data to see
which one applies.

The function saves results in the input directory as csv files with a "," as separator.
One file is settings.csv with all settings, the other one is measures.csv
It saves them directly for not loading all of them at the same time in the session.}
}
\description{
Reading several MicroPEM files and converting the settings and measurement tables to csv.
}

