% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Columbia}
\alias{Columbia}
\title{A  micropem output file.}
\format{A  csv micropem output file of version Columbia of the output file.}
\description{
A  csv micropem output file of version Columbia of the output file.
}
\examples{
\dontrun{
micropemExample <- convert_output(system.file('extdata', 'Columbia.csv', package = 'rtimicropem'))
micropemExample$plot()
}
}
\keyword{data}
