% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_output.R
\name{convert_output}
\alias{convert_output}
\title{Uses output file from MicroPEM to create a \code{micropem} object.}
\usage{
convert_output(path)
}
\arguments{
\item{path}{the path to the file (\code{character})}
}
\value{
A \code{micropem} object.
}
\description{
Uses output file from MicroPEM to create a \code{micropem} object.
}
\examples{
micropem_example <- convert_output(system.file('extdata', 'CHAI.csv', package = 'rtimicropem'))
micropem_example$plot()
micropem_example$plot(title = "wow")
}
