% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micropem.R
\docType{package}
\name{micropem-package}
\alias{micropem}
\alias{micropem-package}
\title{Supports the analyses of RTI MicroPEM output files.}
\description{
micropem allows you to read and analyse data from MicroPEM output files
in a reproducible way.
}

