% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{micropemChai}
\alias{micropemChai}
\title{A  micropem object.}
\format{A  micropem object.}
\description{
A  micropem object.
}
\examples{
data("micropemChai")
micropemChai$plot()
}
\keyword{data}
