% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micropem.R
\docType{package}
\name{rtimicropem-package}
\alias{rtimicropem}
\alias{rtimicropem-package}
\title{Supports the analyses of RTI MicroPEM output files.}
\description{
micropem allows you to read and analyse data from MicroPEM output files
in a reproducible way.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/rtimicropem}
  \item \url{https://docs.ropensci.org/rtimicropem/}
  \item Report bugs at \url{https://github.com/ropensci/rtimicropem/issues}
}

}
\author{
\strong{Maintainer}: Maëlle Salmon \email{maelle.salmon@yahoo.se}

Authors:
\itemize{
  \item Zheng Zhou (Department of Environmental Health Sciences Mailman School of Public Health at Columbia University, New York, USA)
}

Other contributors:
\itemize{
  \item  ERC Grant Agreement number 336167 - the CHAI Project [funder]
  \item Carles Milà [contributor]
  \item Sreekanth Vakacherla [contributor]
  \item Cathryn Tonne [contributor]
  \item Julian Marshall [contributor]
  \item Lucy D'Agostino McGowan (Lucy D'Agostino McGowan reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/126) [reviewer]
  \item Kara  Woo (Kara Woo reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/126) [reviewer]
}

}
