% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run a built-in Shiny App.}
\usage{
run_shiny_app()
}
\description{
Run a built-in Shiny App.
}
\details{
This app allows you to upload a single MicroPEM file and to explore it via several tabs,
one with the summary of measures, one with settings, one with a plot and with alarms (see \code{\link[rtimicropem]{chai_alarm}}).
The app can e.g. be used after one day of field work to quickly check files.
}
\examples{
\dontrun{
run_shiny_app()
}
}
